+++ oedatamodel_api/settings.py
@@ -1,7 +1,7 @@
 import os
 import pathlib
 
-VERSION = "0.8.0"
+VERSION = "0.9.0"
 
 ROOT_DIR = pathlib.Path(__file__).resolve().parent.parent
 APP_DIR = ROOT_DIR / "oedatamodel_api"