+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v
 ### Added
 - query page with dynamic source parameters
 - project-specific mappings and sources
+- "table" mapping for raw data tables from OEP
 
 ## [0.8.0] - 2022-06-15
 ### Added