+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v
 - query page with dynamic source parameters
 - project-specific mappings and sources
 - "table" mapping for raw data tables from OEP
+- support for single csv output format
 
 ## [0.8.0] - 2022-06-15
 ### Added